Extract GTModelDescriptor test method to class
